requirements.txt
streamlit # add other python packages that you need google to install in order for your app to run
google-cloud-aiplatform #this is required
google-cloud-logging  #this is required
matplotlib
numpy
mysql.connector  # Assuming you're using MySQL for GCP
pandas
altair
git+https://github.com/pandas-profiling/pandas-profiling@master
streamlit_pandas_profiling
sweetviz
pydantic-settings